Reorganize a few remaining files
