+++ TeamCode/src/main/java/com/acmerobotics/relicrecovery/localization/CryptoboxLocalizer.java
@@ -1,4 +1,4 @@
-package com.acmerobotics.relicrecovery.vision;
+package com.acmerobotics.relicrecovery.localization;
 
 import android.util.Log;
 
@@ -8,6 +8,8 @@ import com.acmerobotics.library.localization.Vector2d;
 import com.acmerobotics.relicrecovery.configuration.AllianceColor;
 import com.acmerobotics.relicrecovery.configuration.Cryptobox;
 import com.acmerobotics.relicrecovery.subsystems.MecanumDrive;
+import com.acmerobotics.relicrecovery.vision.CryptoboxTracker;
+import com.acmerobotics.relicrecovery.vision.VisionCamera;
 
 import java.util.ArrayList;
 import java.util.Collections;
