+++ TeamCode/src/main/java/com/acmerobotics/relicrecovery/opmodes/test/AutoPathTest.java
@@ -5,7 +5,7 @@ import com.acmerobotics.library.dashboard.canvas.Canvas;
 import com.acmerobotics.library.localization.Pose2d;
 import com.acmerobotics.relicrecovery.configuration.BalancingStone;
 import com.acmerobotics.relicrecovery.opmodes.AutoPaths;
-import com.acmerobotics.relicrecovery.util.DrawingUtil;
+import com.acmerobotics.library.util.DrawingUtil;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
