+++ TeamCode/src/main/java/com/acmerobotics/relicrecovery/opmodes/test/CryptoPositionTest.java
@@ -7,8 +7,8 @@ import com.acmerobotics.library.localization.Vector2d;
 import com.acmerobotics.relicrecovery.configuration.AllianceColor;
 import com.acmerobotics.relicrecovery.subsystems.MecanumDrive;
 import com.acmerobotics.relicrecovery.subsystems.Robot;
-import com.acmerobotics.relicrecovery.util.DrawingUtil;
-import com.acmerobotics.relicrecovery.vision.CryptoboxLocalizer;
+import com.acmerobotics.library.util.DrawingUtil;
+import com.acmerobotics.relicrecovery.localization.CryptoboxLocalizer;
 import com.acmerobotics.relicrecovery.vision.CryptoboxTracker;
 import com.acmerobotics.relicrecovery.vision.VuforiaCamera;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
